+++ src/Features/Core/Portable/GenerateMember/GenerateParameterizedMember/AbstractGenerateDeconstructMethodService.State.cs
@@ -10,7 +10,6 @@ using System.Threading.Tasks;
 using System.Xml.Linq;
 using Microsoft.CodeAnalysis.CodeGeneration;
 using Microsoft.CodeAnalysis.CodeRefactorings;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Editing;
 using Microsoft.CodeAnalysis.LanguageService;
 using Microsoft.CodeAnalysis.PooledObjects;
@@ -70,7 +69,7 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateParameterizedMember
                     return false;
                 }
 
-                var parameters = TryMakeParameters(semanticModel, targetVariables, semanticFacts, cancellationToken);
+                var parameters = service.TryMakeParameters(semanticModel, targetVariables, semanticFacts, cancellationToken);
                 if (parameters.IsDefault)
                 {
                     return false;
@@ -90,47 +89,6 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateParameterizedMember
                 SignatureInfo = new MethodSignatureInfo(document, this, methodSymbol);
 
                 return await TryFinishInitializingStateAsync(service, document, cancellationToken).ConfigureAwait(false);
-            }
-
-            private static ImmutableArray<IParameterSymbol> TryMakeParameters(SemanticModel semanticModel, SyntaxNode target, ISemanticFactsService semanticFacts, CancellationToken cancellationToken)
-            {
-                ITypeSymbol targetType;
-                if (target is PositionalPatternClauseSyntax positionalPattern)
-                {
-                    var namesBuilder = ImmutableArray.CreateBuilder<string>();
-                    using var _ = ArrayBuilder<IParameterSymbol>.GetInstance(positionalPattern.Subpatterns.Count, out var builder);
-                    for (var i = 0; i < positionalPattern.Subpatterns.Count; i++)
-                    {
-                        namesBuilder.Add(semanticFacts.GenerateNameForExpression(semanticModel, ((ConstantPatternSyntax)positionalPattern.Subpatterns[i].Pattern).Expression, false, cancellationToken));
-                    }
-                    var names = NameGenerator.EnsureUniqueness(namesBuilder.ToImmutable());
-                    for (var i = 0; i < positionalPattern.Subpatterns.Count; i++)
-                    {
-                        targetType = semanticModel.GetTypeInfo(((ConstantPatternSyntax)positionalPattern.Subpatterns[i].Pattern).Expression, cancellationToken: cancellationToken).Type;
-                        builder.Add(CodeGenerationSymbolFactory.CreateParameterSymbol(
-                        attributes: default, RefKind.Out, isParams: false, targetType, names[i]));
-                    }
-
-                    return builder.ToImmutableAndClear();
-                }
-                else
-                {
-                    targetType = semanticModel.GetTypeInfo(target, cancellationToken: cancellationToken).Type;
-                    if (targetType?.IsTupleType != true)
-                    {
-                        return default;
-                    }
-
-                    var tupleElements = ((INamedTypeSymbol)targetType).TupleElements;
-                    using var _ = ArrayBuilder<IParameterSymbol>.GetInstance(tupleElements.Length, out var builder);
-                    foreach (var element in tupleElements)
-                    {
-                        builder.Add(CodeGenerationSymbolFactory.CreateParameterSymbol(
-                            attributes: default, RefKind.Out, isParams: false, element.Type, element.Name));
-                    }
-
-                    return builder.ToImmutableAndClear();
-                }
             }
         }
     }
